+++ pkg/k8sclient/awsclusterclient.go
@@ -8,6 +8,7 @@ import (
 	capa "sigs.k8s.io/cluster-api-provider-aws/api/v1beta1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/util"
+	"sigs.k8s.io/cluster-api/util/conditions"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
@@ -65,4 +66,10 @@ func (a *AWSClusterClient) GetIdentity(ctx context.Context, awsCluster *capa.AWS
 	}
 
 	return roleIdentity, nil
+}
+
+func (a *AWSClusterClient) MarkConditionTrue(ctx context.Context, awsCluster *capa.AWSCluster, condition capi.ConditionType) error {
+	originalCluster := awsCluster.DeepCopy()
+	conditions.MarkTrue(awsCluster, condition)
+	return a.client.Status().Patch(ctx, awsCluster, client.MergeFrom(originalCluster))
 }